計算ロジックをモデルに移動
