+++ serverside_challenge_1/challenges/sonoki/app/controllers/api/v1/costs_controller.rb
@@ -1,5 +1,5 @@
 class Api::V1::CostsController < ApplicationController
-  require 'yaml'
+  include YamlLoader
 
   def index
     render json: {
@@ -9,49 +9,20 @@ class Api::V1::CostsController < ApplicationController
     }, status: 200
   end
 
-  def calculate_rate # 10行で収める その他はモデルに移動
-    contract_ampere = params[:contract_ampere].to_i
-    usage = params[:usage].to_i unless params[:usage].nil?
-    rates = yaml_data
-
-    if contract_ampere.zero? || usage.nil?
-      render json: { error: 'Invalid input: contract_ampere and usage are required' }, status: 400
+  def calculate_rate
+    @cost_query = Cost.new(query_params)
+    if @cost_query.valid?
+      @costs = @cost_query.calculate
+      render json: @costs, status: 200
     else
-      costs = []
-      rates.each do |key, values|
-        basic_rate = values['basic_rates'][contract_ampere]
-        next unless basic_rate
-
-        usage_total_cost = 0
-        remaining_usage = usage
-
-        values['usage_rates'].each do |rate_info|
-          min, max = rate_info['range'].map { |value| value == "inf" ? Float::INFINITY : value }
-          difference =  max - min + 1
-          if remaining_usage >= difference
-            usage_total_cost += rate_info['rate'] * difference
-            remaining_usage -= difference
-          else
-            usage_total_cost += rate_info['rate'] * remaining_usage
-            remaining_usage -= remaining_usage
-            break
-          end
-        end
-        total_cost = basic_rate + usage_total_cost
-
-        costs << { provider_name: key,
-                    plan_name: values['plan_name'],
-                    price: total_cost
-                  }
-      end
-      render json: costs, status: 200
+      render json: { error: 'Invalid input: contract_ampere and usage are required' }, status: 400
     end
   end
 
   private
 
-  def yaml_data
-    yaml_data ||= YAML.load_file(Rails.root.join('config', 'rates.yml'))
+  def query_params
+    params.permit(:contract_ampere, :usage).to_h.with_indifferent_access
   end
 
 end
