+++ serverside_challenge_1/challenges/sonoki/app/controllers/application_controller.rb
@@ -1,2 +1,7 @@
 class ApplicationController < ActionController::API
+  # require 'yaml'
+
+  # def yaml_data
+  #   @yaml_data ||= YAML.load_file(Rails.root.join('config', 'rates.yml'))
+  # end
 end
