+++ serverside_challenge_1/challenges/sonoki/app/lib/yaml_loader.rb
@@ -1,0 +1,6 @@
+module YamlLoader
+  def yaml_data
+    require 'yaml'
+    @yaml_data ||= YAML.load_file(Rails.root.join('config', 'rates.yml'))
+  end
+end
