+++ serverside_challenge_1/challenges/sonoki/config/application.rb
@@ -24,6 +24,9 @@ module Sonoki
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.1
 
+    # lib ディレクトリを autoload_paths に追加
+    config.autoload_paths += %W(#{config.root}/app/lib)
+
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
